ARCH-1: extract types/lint-rules dispatch arms (iter-225)
