+++ crates/hyalo-cli/src/commands/lint_rules.rs
@@ -930,3 +930,70 @@ mod tests {
         assert!(activation_for("HYALO001", &empty).is_none());
     }
 }
+
+// ---------------------------------------------------------------------------
+// Dispatch handler (ARCH-1, iter-225)
+// ---------------------------------------------------------------------------
+
+/// The `hyalo lint-rules` dispatch arm, extracted verbatim from `dispatch.rs`.
+pub(crate) fn run(
+    ctx: &mut crate::dispatch::CommandContext<'_>,
+    action: Option<crate::cli::args::LintRulesAction>,
+) -> Result<CommandOutcome> {
+    let effective_format = ctx.effective_format;
+
+
+    let action = action.unwrap_or(crate::cli::args::LintRulesAction::List {
+        enabled_only: false,
+        disabled_only: false,
+        rule_prefix: None,
+    });
+    let md_engine = hyalo_mdlint::HyaloLintEngine::create()
+        .map_err(|e| anyhow::anyhow!("failed to create lint engine: {e}"))?;
+    match action {
+        crate::cli::args::LintRulesAction::List {
+            enabled_only,
+            disabled_only,
+            rule_prefix,
+        } => Ok(crate::commands::lint_rules::list_rules(
+            ctx.config_dir,
+            &md_engine,
+            ctx.md_lint,
+            ctx.schema,
+            enabled_only,
+            disabled_only,
+            rule_prefix.as_deref(),
+            effective_format,
+        )),
+        crate::cli::args::LintRulesAction::Show { rule_id } => Ok(crate::commands::lint_rules::show_rule(
+            &rule_id,
+            &md_engine,
+            ctx.md_lint,
+            ctx.schema,
+            ctx.user_format,
+        )),
+        crate::cli::args::LintRulesAction::Set {
+            rule_id,
+            enabled,
+            severity,
+            dry_run,
+        } => crate::commands::lint_rules::set_rule(
+            ctx.config_dir,
+            &rule_id,
+            enabled,
+            severity.as_deref(),
+            dry_run,
+            &md_engine,
+            ctx.md_lint,
+            ctx.user_format,
+        ),
+        crate::cli::args::LintRulesAction::Remove { rule_id, dry_run } => crate::commands::lint_rules::remove_rule(
+            ctx.config_dir,
+            &rule_id,
+            dry_run,
+            &md_engine,
+            ctx.md_lint,
+            ctx.user_format,
+        ),
+    }
+}
